# fix/perf: optimize splits by caching expensive repeated calls