--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -37,10 +37,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- Optimize dynamic splits by caching repeated calls within callback.
 - CoinGecko pricing source support
 - Consistently rename min liqudity filter across pool models & configs to min pool liquidity capitalization (cap)
 - Validate token in and out in /router quote endpoints to ensure that they are not equal to each other & clean up tokens/prices
 - Hack to fallback to precision of 18 when estimating spot price for Astroport pools via quotes.
+
+## 0.18.5
+
+- Asset list V2 integration
 
 ## 0.18.4