… router/usecase/routertesting/parsing/taker_fees.json
… router/usecase/routertesting/parsing/tokens.json