While working on https://github.com/osmosis-labs/sqs/pull/268 (PEPE splits bug), I noticed that a performance optimization from before was not functioning correctly due to misues of closures.

The performance optimization is intended to cache repeated quote estimation calls in the dynamic programming solution within closure scope to avoid expensive recomputation

This PR fixes the bug, and adds a sanity check test and benchmark.

According to the benchmark, this introduces an 80% speed up to splits and reduces reallocations by the same magnitude.

```
goos: darwin
goarch: arm64
pkg: github.com/osmosis-labs/sqs/router/usecase
                 │  bench_old  │              bench_new              │
                 │   sec/op    │   sec/op     vs base                │
GetSplitQuote-11   814.4µ ± 2%   159.0µ ± 1%  -80.47% (p=0.000 n=10)

                 │  bench_old   │              bench_new               │
                 │     B/op     │     B/op      vs base                │
GetSplitQuote-11   817.6Ki ± 0%   155.1Ki ± 0%  -81.03% (p=0.000 n=10)

                 │  bench_old   │              bench_new              │
                 │  allocs/op   │  allocs/op   vs base                │
GetSplitQuote-11   22.724k ± 0%   4.336k ± 0%  -80.92% (p=0.000 n=10)
```

## Drive By

I also regenerated mainnet test state